extract deeplink bridge logic into its own file
